refactor states and functions to improve code structure and optimize for components to be added
